TOOLS: move `CmdCompleter` to `tools-repl`
