+++ cli/src/main/java/org/aya/cli/repl/Repl.java
@@ -8,8 +8,8 @@ import org.aya.api.distill.DistillerOptions;
 import org.aya.api.error.Problem;
 import org.aya.api.util.InterruptException;
 import org.aya.api.util.NormalizeMode;
+import org.aya.cli.repl.jline.AyaCompleters;
 import org.aya.cli.repl.jline.JlineRepl;
-import org.aya.cli.repl.jline.completer.AyaCompleters;
 import org.aya.cli.single.CliReporter;
 import org.aya.cli.utils.MainArgs;
 import org.aya.prelude.GeneratedVersion;
