+++ cli/src/main/java/org/aya/cli/repl/jline/AyaCompleters.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2021 Yinsen (Tesla) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.repl.jline.completer;
+package org.aya.cli.repl.jline;
 
 import kala.tuple.Tuple;
 import kala.tuple.Tuple2;
